funcionario

+++ dadosSpringJPA/src/main/java/pos/dados/entity/Departamento.java
@@ -17,12 +17,12 @@ public class Departamento {
 		super();
 	}	
 	
-	public Departamento(Long cod_departamento) {
+	public Departamento(Integer cod_departamento) {
 		super();
 		this.cod_departamento = cod_departamento;
 	}
 	
-	public Departamento(Long cod_departamento, String nome_departamento) {
+	public Departamento(Integer cod_departamento, String nome_departamento) {
 		super();
 		this.cod_departamento = cod_departamento;
 		this.nome_departamento = nome_departamento;
@@ -30,13 +30,14 @@ public class Departamento {
 	
 	@Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private Long cod_departamento;
+    private Integer cod_departamento;
 	private String nome_departamento;
-	public Long getCod_departamento() {
+	
+	public Integer getCod_departamento() {
 		return cod_departamento;
 	}
 
-	public void setCod_departamento(Long cod_departamento) {
+	public void setCod_departamento(Integer cod_departamento) {
 		this.cod_departamento = cod_departamento;
 	}
